# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visitors:
    post:
      operationId: createVisitor
      tags: [visitors]
      summary: Create a passport visitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVisitorInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
  /visitors/link:
    post:
      operationId: linkVisitor
      tags: [visitors]
      summary: Get or create the passport visitor for the signed-in account
      description: >
        Resolves the visitor linked to the authenticated Clerk session. Creates
        one from the account's profile on first sign-in. Requires a valid session.
      responses:
        "200":
          description: Linked visitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
        "401":
          description: Not authenticated
  /visitors/{id}:
    get:
      operationId: getVisitor
      tags: [visitors]
      summary: Get a visitor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Visitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
        "404":
          description: Not found
...
  /visitors/{id}/preferences:
    patch:
      operationId: updateVisitorPreferences
      tags: [visitors]
      summary: Update a visitor's communication preferences (promo email consent)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVisitorPreferencesInput"
      responses:
        "200":
          description: Updated visitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
  /admin/auth/login:
    post:
      operationId: staffLogin
      tags: [admin]
      summary: Staff login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginInput"
      responses:
        "200":
          description: Logged in (session cookie set)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSessionInfo"
...
  /admin/auth/change-password:
    post:
      operationId: staffChangePassword
      tags: [admin]
      summary: Change the current staff user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSessionInfo"
...
  /admin/users/{id}:
    patch:
      operationId: updateStaffUser
      tags: [admin]
      summary: Update a staff user's email or status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
  /admin/audit-log:
    get:
      operationId: listAdminAuditLog
      tags: [admin]
      summary: Recent admin action history
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAuditEntry"
  /admin/settings/legacy-access:
    get:
      operationId: getLegacyAccessStatus
      tags: [admin]
      summary: Whether the legacy shared admin secret is still accepted
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyAccessStatus"
    patch:
      operationId: setLegacyAccessStatus
      tags: [admin]
      summary: Enable or disable the legacy shared admin secret (staff session required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyAccessStatus"
      responses:
        "200":
          description: Updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyAccessStatus"
...
components:
  schemas:
...
    Visitor:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        promoOptIn:
          type: boolean
        promoOptInAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, firstName, email, promoOptIn, createdAt]
    CreateVisitorInput:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
      required: [firstName, email]
...
    StaffLoginInput:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
      required: [username, password]
    StaffChangePasswordInput:
      type: object
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 10 }
      required: [currentPassword, newPassword]
    StaffUser:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, nullable: true }
        status: { type: string }
        mustChangePassword: { type: boolean }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        passwordChangedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, username, status, mustChangePassword, createdAt]
    StaffSessionInfo:
      type: object
      properties:
        authenticated: { type: boolean }
        user:
          $ref: "#/components/schemas/StaffUser"
      required: [authenticated]
    UpdateStaffUserInput:
      type: object
      properties:
        email: { type: string, nullable: true }
        status:
          type: string
          enum: [active, disabled]
    AdminAuditEntry:
      type: object
      properties:
        id: { type: string }
        actor: { type: string }
        action: { type: string }
        entityType: { type: string, nullable: true }
        entityId: { type: string, nullable: true }
        detail: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, actor, action, createdAt]
    LegacyAccessStatus:
      type: object
      properties:
        enabled: { type: boolean }
      required: [enabled]
    UpdateVisitorPreferencesInput:
      type: object
      properties:
        promoOptIn: { type: boolean }
      required: [promoOptIn]
...