# lib/api-zod/src/generated/types/adminAuditEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAuditEntry {
  id: string;
  actor: string;
  action: string;
  entityType?: string | null;
  entityId?: string | null;
  detail?: string | null;
  createdAt: Date;
}